application/app_service/pyproject.toml
[tool.poetry]
name = "postbin"
version = "0.1.0"
description = ""
authors = ["Vadim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
uvicorn = "^0.32.1"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
bcrypt = "^4.2.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
cryptography = "43.0.3"
pika = "^1.3.2"
python-dotenv = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.10.3"}
aiobotocore = "^2.15.2"
aiofiles = "^24.1.0"
python-multipart = "^0.0.20"
prometheus-fastapi-instrumentator = "^7.0.0"
aioredis = "^2.0.1"
pytest = "^8.3.4"
factory-boy = "^3.3.1"
pytest-factoryboy = "^2.7.0"
pytest-asyncio = "^0.25.2"
trio = "^0.28.0"
httpx = "^0.28.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
